wip: routing + some refactors + env config
