+++ package.json
@@ -14,7 +14,7 @@
         "lucide-react": "^0.406.0",
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
-        "react-router-dom": "^7.2.0"
+        "react-router-dom": "^7.5.3"
     },
     "devDependencies": {
         "@vitejs/plugin-react": "^4.3.4",
