+++ src/constants.js
@@ -1,0 +1,7 @@
+
+export const CollectionNames = {
+  DATES: "dates",
+  REGISTRATIONS: "registrations",
+  ADMINS: "admins",
+  MAIL: "mail",
+};
