+++ src/main.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import ListingApp from './components/ListingApp';
+import Routes from './Routes';
 import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
     <React.StrictMode>
-        <ListingApp />
+        <Routes />
     </React.StrictMode>
 );
